fix: compress assets outside of rollup
resolves: nuxt/framework#7197
